refactor: LadderGenerationService 클래스와 LadderGameService 클래스 통합  - LadderGameService 클래스로 기능 통합
